fix: resolve test compilation issues and enhance CI setup
- Fix unused variable and import issues in security tests
- Add proper error handling in benchmark tests
- Enhanced command injection test coverage
- Fix type conversion issues in integration tests
- Add Kubebuilder setup to CI workflow for controller testing

These changes ensure all tests compile and run successfully while maintaining
comprehensive security test coverage.

+++ tests/security/integration_security_test.go
@@ -758,7 +758,7 @@ func (suite *IntegrationSecurityTestSuite) testPerformanceMetricsUnderAttack(t *
 	attackLoad := suite.simulateAttackLoad(t)
 
 	// Compare performance
-	degradation := (attackLoad - baseline) / baseline
+	degradation := float64(attackLoad-baseline) / float64(baseline)
 	t.Logf("Performance degradation under attack: %.2f%% (baseline: %v, under attack: %v)",
 		degradation*100, baseline, attackLoad)
